removed ScaleNorm. merged files
